UI Snapshots, UI retained mode
- The big feature that was implemented was in the UI - windows can take snapshots of their output and have it reused across frames instead of having to calculate them each frame. The UI windows can now specify a retained function that returns true when the window can reuse the snapshot, else false. The drawing of the windows had to be improved to allow such functionality.
- Now the UI handlers have a separate allocator that can be used to make allocations for buffers easier than embedding the data into the same data source. A new entry was added to the UIHandler - a copy function that is called when a new handler is selected to copy the temporary buffer into that allocator
- More SoA utility functions
- The entity hierarchy hash function was changed to basically identity since it seems to behave better for the time being (that needs to be investigated as the inputs in general monotonic integers that can result in long chains that will make the table grow in case a collision happens on that chain)
- The serialization/deserialization of the EntityManager was changed to accept option structures for easier maintability
- The ForEachEntity implementation was changed to reduce the total amount of bytes used by the ForEachEntityBatchImplementationTaskData since it can result in having to make the task allocators for the threads large in case there are many batches
- Changed selected entities runtime resource to use malloc in case the system allocator cannot handle it
- Added TaskParallelFor for easier addition of dynamic tasks that use fork - join
- The task scheduler no longer uses the barrier since that was moved to the task manager static tasks
- Added some accessor functions for the ResourceManager
- The UIDrawer now has record_actions and record_snapshots boolean flags to help reduce the workfload. Now IsMouseInRectangle function is removed since it can result in unwanted results since that will not be included in snapshots. For that you need to use SnapshotRunnable or the classic hoverable/clickable/general handlers
- ForEach File/Directory functions now have templated lambda wrappers to allow easier captures
- The Editor FileExplorer is the first window to implement retained mode (there are still some modifications to be made to catch all cases)

+++ ECSEngine/src/ECSEngine/Utilities/FilePackaging.cpp
@@ -188,7 +188,7 @@ namespace ECSEngine {
 		memset(&table, 0, sizeof(table));
 
 		// Read the last 4 bytes - the size of the table
-		size_t status = SetFileCursor(file_handle, -sizeof(TableHeader), ECS_FILE_SEEK_END);
+		size_t status = SetFileCursor(file_handle, -(int64_t)sizeof(TableHeader), ECS_FILE_SEEK_END);
 
 		// Failed to seek
 		if (status == -1) {
@@ -213,7 +213,7 @@ namespace ECSEngine {
 
 		// Now read the whole data into a buffer - the string and the actual table
 		void* allocation = AllocateEx(allocator, header.table_size + header.string_total_size);
-		status = SetFileCursor(file_handle, -sizeof(header) - header.table_size - header.string_total_size, ECS_FILE_SEEK_END);
+		status = SetFileCursor(file_handle, -(int64_t)sizeof(header) - header.table_size - header.string_total_size, ECS_FILE_SEEK_END);
 		if (status == -1) {
 			DeallocateEx(allocator, allocation);
 			return table;
